refactor: move agent command builder to shared controller package
